+++ build_runner/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.1.5
+
+- Allow the latest `build_daemon`.
+
 ## 2.1.4
 
 - Minor fixes to the log message when compiling the build script has some